fix load error if russ.db.new doesnt exist

+++ src/cash/load.sh
@@ -3,5 +3,5 @@
 set -e
 set -x
 
-rm -iv russ.db.new
+test ! -e russ.db.new || rm -iv russ.db.new
 sqlite3 russ.db.new < russ.db.sql
